requirements.txt
Flask==1.1.2
python-dotenv==0.12.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
Flask-WTF==0.14.3
email-validator==1.0.5